+++ src/main/java/be/kdg/programming5/programming5/domain/MenuItemChef.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
 
 @Entity
 @Table(uniqueConstraints = { @UniqueConstraint(columnNames = { "menu_item_id", "chef_id" }) })
-public class MenuItemChef extends AbstractEntity<Integer> implements Serializable {
+public class MenuItemChef extends AbstractEntity<Long> implements Serializable {
     @ManyToOne(optional = false, fetch = FetchType.LAZY)
     @JoinColumn(name = "menu_item_id")
     private MenuItem menuItem;
@@ -19,7 +19,7 @@ public class MenuItemChef extends AbstractEntity<Integer> implements Serializabl
     public MenuItemChef() {
     }
 
-    public MenuItemChef(int id, MenuItem menuItem, Chef chef) {
+    public MenuItemChef(long id, MenuItem menuItem, Chef chef) {
         super(id);
         this.menuItem = menuItem;
         this.chef = chef;
