+++ src/main/java/be/kdg/programming5/programming5/repository/MenuItemChefRepository.java
@@ -3,5 +3,5 @@ package be.kdg.programming5.programming5.repository;
 import be.kdg.programming5.programming5.domain.MenuItemChef;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuItemChefRepository extends JpaRepository<MenuItemChef, Integer> {
+public interface MenuItemChefRepository extends JpaRepository<MenuItemChef, Long> {
 }
